requirements.txt
python-dotenv==1.0.0
spacy==3.7.2
textblob==0.17.1
crewai==0.51.1 
docstring-parser==0.16
instructor==1.3.3 
langchain==0.2.14 
langchain-community==0.2.12
langchain-core==0.2.35
langchain-openai==0.1.22
langchain-text-splitters==0.2.2
langsmith==0.1.104 
openai==1.42.0 
tiktoken==0.7.0
duckduckgo-search==6.2.11
primp==0.6.1